Some PR cleanups

+++ filesystem/src/main/java/org/frankframework/filesystem/smb/Samba1FileSystem.java
@@ -21,8 +21,8 @@ import java.nio.file.DirectoryStream;
 import java.util.Date;
 import java.util.Iterator;
 import java.util.Map;
+import java.util.NoSuchElementException;
 
-import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 import jcifs.smb.NtlmPasswordAuthentication;
@@ -34,7 +34,9 @@ import jcifs.smb.SmbFileOutputStream;
 import lombok.Getter;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.logging.log4j.util.Supplier;
 import org.frankframework.configuration.ConfigurationException;
+import org.frankframework.filesystem.FileAlreadyExistsException;
 import org.frankframework.filesystem.FileNotFoundException;
 import org.frankframework.filesystem.FileSystemBase;
 import org.frankframework.filesystem.FileSystemException;
@@ -103,7 +105,7 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	}
 
 	@Override
-	public SmbFile toFile(@Nullable String folder, @Nonnull @Nullable String filename) throws FileSystemException {
+	public SmbFile toFile(@Nullable String folder, @Nullable String filename) throws FileSystemException {
 		return toFile(folder+"/"+filename);
 	}
 
@@ -213,6 +215,12 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 
 	@Override
 	public SmbFile renameFile(SmbFile source, SmbFile destination) throws FileSystemException {
+		if (!exists(source)) {
+			throw new FileNotFoundException("Cannot find file [" + getName(source) + "], cannot rename.");
+		}
+		if (exists(destination)) {
+			throw new FileAlreadyExistsException("Cannot rename [" + getName(source) + "], destination [" + getName(destination) + "] already exists");
+		}
 		try {
 			source.renameTo(destination);
 			return destination;
@@ -222,7 +230,18 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	}
 
 	@Override
-	public SmbFile moveFile(SmbFile f, String destinationFolder, boolean createFolder, boolean resultantMustBeReturned) throws FileSystemException {
+	public SmbFile moveFile(SmbFile f, String destinationFolder, boolean createFolder) throws FileSystemException {
+		if (!exists(f)) {
+			throw new FileNotFoundException("Cannot find file [" + getName(f) + "], cannot move.");
+		}
+		if (!folderExists(destinationFolder)) {
+			if(createFolder) {
+				createFolder(destinationFolder);
+			} else {
+				throw new FolderNotFoundException("destination does not exist");
+			}
+		}
+
 		SmbFile dest = toFile(destinationFolder, f.getName());
 		try {
 			f.renameTo(dest);
@@ -233,7 +252,10 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 	}
 
 	@Override
-	public SmbFile copyFile(SmbFile f, String destinationFolder, boolean createFolder, boolean resultantMustBeReturned) throws FileSystemException {
+	public SmbFile copyFile(SmbFile f, String destinationFolder, boolean createFolder) throws FileSystemException {
+		if (!exists(f)) {
+			throw new FileNotFoundException("Cannot find file [" + getName(f) + "], cannot copy.");
+		}
 		if (!folderExists(destinationFolder)) {
 			if(createFolder) {
 				createFolder(destinationFolder);
@@ -259,17 +281,11 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 
 	private class SmbFileIterator implements Iterator<SmbFile> {
 
-		private SmbFile[] files;
+		private final SmbFile[] files;
 		private int i = 0;
 
 		public SmbFileIterator(String folder) throws IOException {
-			SmbFileFilter filter = new SmbFileFilter() {
-
-				@Override
-				public boolean accept(SmbFile file) throws SmbException {
-					return (!isListHiddenFiles() && !file.isHidden()) && file.isFile();
-				}
-			};
+			SmbFileFilter filter = file -> (!isListHiddenFiles() && !file.isHidden()) && file.isFile();
 			SmbFile f = smbContext;
 			if(StringUtils.isNotBlank(folder)) {
 				String normalizedFolder = FilenameUtils.normalizeNoEndSeparator(folder, true)+"/";
@@ -285,6 +301,9 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 
 		@Override
 		public SmbFile next() {
+			if (i >= files.length) {
+				throw new NoSuchElementException();
+			}
 			return files[i++];
 		}
 
@@ -294,7 +313,7 @@ public class Samba1FileSystem extends FileSystemBase<SmbFile> implements IWritab
 			try {
 				deleteFile(file);
 			} catch (FileSystemException e) {
-				log.warn("unable to delete file ["+getCanonicalName(file)+"]", e);
+				log.warn("unable to delete file [{}]", (Supplier<?>) ()->getCanonicalName(file), e);
 			}
 		}
 	}
